# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "CheckList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CheckSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "CheckList",
        "type": "object"
      },
      "CheckSummary": {
        "properties": {
          "dataset": {
            "title": "Dataset",
            "type": "string"
          },
          "datasource": {
            "title": "Datasource",
            "type": "string"
          },
          "expression": {
            "title": "Expression",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "unit": {
            "enum": [
              "count",
              "percent",
              "duration",
              "number"
            ],
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "expression",
          "metric",
          "unit",
          "dataset",
          "datasource",
          "owner",
          "tags",
          "source",
          "location",
          "latest"
        ],
        "title": "CheckSummary",
        "type": "object"
      },
      "Counts": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "integer"
          },
          "fail": {
            "title": "Fail",
            "type": "integer"
          },
          "pass": {
            "title": "Pass",
            "type": "integer"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "warn": {
            "title": "Warn",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "pass",
          "warn",
          "fail",
          "error",
          "skipped"
        ],
        "title": "Counts",
        "type": "object"
      },
      "Datasource": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "Datasource",
        "type": "object"
      },
      "Diagnostic": {
        "properties": {
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "enum": [
              "error",
              "warning"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "message",
          "location"
        ],
        "title": "Diagnostic",
        "type": "object"
      },
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorInfo": {
        "properties": {
          "code": {
            "enum": [
              "invalid_parameter",
              "forbidden_host",
              "not_found",
              "method_not_allowed",
              "internal_error",
              "store_unavailable"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo",
        "type": "object"
      },
      "HistoryEntry": {
        "properties": {
          "display_value": {
            "title": "Display Value",
            "type": "string"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "number"
          },
          "expression": {
            "title": "Expression",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "pass",
              "warn",
              "fail",
              "error",
              "skipped"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "run_id",
          "started_at",
          "trigger",
          "outcome",
          "value",
          "display_value",
          "message",
          "duration_ms",
          "name",
          "expression",
          "source"
        ],
        "title": "HistoryEntry",
        "type": "object"
      },
      "HistoryPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryEntry"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "HistoryPage",
        "type": "object"
      },
      "LatestResult": {
        "properties": {
          "display_value": {
            "title": "Display Value",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "outcome": {
            "enum": [
              "pass",
              "warn",
              "fail",
              "error",
              "skipped"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "run_id",
          "started_at",
          "trigger",
          "outcome",
          "value",
          "display_value",
          "message"
        ],
        "title": "LatestResult",
        "type": "object"
      },
      "Location": {
        "properties": {
          "column": {
            "title": "Column",
            "type": "integer"
          },
          "file": {
            "title": "File",
            "type": "string"
          },
          "line": {
            "title": "Line",
            "type": "integer"
          }
        },
        "required": [
          "file",
          "line",
          "column"
        ],
        "title": "Location",
        "type": "object"
      },
      "Project": {
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/ProjectCounts"
          },
          "datasources": {
            "items": {
              "$ref": "#/components/schemas/Datasource"
            },
            "title": "Datasources",
            "type": "array"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "loaded_at": {
            "title": "Loaded At",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "version",
          "loaded_at",
          "ok",
          "datasources",
          "counts",
          "diagnostics"
        ],
        "title": "Project",
        "type": "object"
      },
      "ProjectCounts": {
        "properties": {
          "checks": {
            "title": "Checks",
            "type": "integer"
          },
          "datasets": {
            "title": "Datasets",
            "type": "integer"
          }
        },
        "required": [
          "datasets",
          "checks"
        ],
        "title": "ProjectCounts",
        "type": "object"
      },
      "Run": {
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/Counts"
          },
          "exit_code": {
            "title": "Exit Code",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "pass",
              "warn",
              "fail",
              "error",
              "skipped"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "project": {
            "title": "Project",
            "type": "string"
          },
          "selection": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Selection",
            "type": "object"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project",
          "started_at",
          "finished_at",
          "outcome",
          "exit_code",
          "trigger",
          "version",
          "selection",
          "counts"
        ],
        "title": "Run",
        "type": "object"
      },
      "RunDetail": {
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/Counts"
          },
          "exit_code": {
            "title": "Exit Code",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "pass",
              "warn",
              "fail",
              "error",
              "skipped"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "project": {
            "title": "Project",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RunResultItem"
            },
            "title": "Results",
            "type": "array"
          },
          "selection": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Selection",
            "type": "object"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project",
          "started_at",
          "finished_at",
          "outcome",
          "exit_code",
          "trigger",
          "version",
          "selection",
          "counts",
          "results"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Run"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "RunPage",
        "type": "object"
      },
      "RunResultItem": {
        "properties": {
          "check_id": {
            "title": "Check Id",
            "type": "string"
          },
          "dataset": {
            "title": "Dataset",
            "type": "string"
          },
          "datasource": {
            "title": "Datasource",
            "type": "string"
          },
          "display_value": {
            "title": "Display Value",
            "type": "string"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "number"
          },
          "expression": {
            "title": "Expression",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "pass",
              "warn",
              "fail",
              "error",
              "skipped"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "check_id",
          "name",
          "expression",
          "metric",
          "dataset",
          "datasource",
          "outcome",
          "value",
          "display_value",
          "message",
          "source",
          "owner",
          "tags",
          "duration_ms"
        ],
        "title": "RunResultItem",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "tablewatch",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/checks": {
      "get": {
        "operationId": "list_checks_api_v1_checks_get",
        "parameters": [
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "pass",
                      "warn",
                      "fail",
                      "error",
                      "skipped",
                      "not_run"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid parameter"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Results store unavailable"
          }
        },
        "summary": "List Checks"
      }
    },
    "/api/v1/checks/{check_id}": {
      "get": {
        "operationId": "get_check_api_v1_checks__check_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "check_id",
            "required": true,
            "schema": {
              "title": "Check Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid parameter"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Results store unavailable"
          }
        },
        "summary": "Get Check"
      }
    },
    "/api/v1/checks/{check_id}/history": {
      "get": {
        "operationId": "check_history_api_v1_checks__check_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "check_id",
            "required": true,
            "schema": {
              "title": "Check Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid parameter"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Results store unavailable"
          }
        },
        "summary": "Check History"
      }
    },
    "/api/v1/project": {
      "get": {
        "operationId": "get_project_api_v1_project_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid parameter"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Results store unavailable"
          }
        },
        "summary": "Get Project"
      }
    },
    "/api/v1/runs": {
      "get": {
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid parameter"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Results store unavailable"
          }
        },
        "summary": "List Runs"
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid parameter"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Results store unavailable"
          }
        },
        "summary": "Get Run"
      }
    }
  }
}